Undo the modification to allow declaring a model more than once
The purpose of this modification of Apps.register_model() was to allow
overriding models of an overridable app by declaring the new model in
an overriding app and setting the overriding app before the overridable
app in INSTALLED_APPS.

The modification was to register only the first model found and ignore
the rest. However, that approach fails since related fields create
reverse fields even if their model is not registered. So, reverse
fields point to the last model found and not to the first, which is
the desirable.